+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated public let properties of public structs with memberwise initializers to be public var.
 - `BarStackView` now handles selection of bar models and can be used as an `EpoxyableView`.
+- The cases of `BarStackView.ZOrder` have been renamed to be more semantically accurate
 
 ## [0.1.0](https://github.com/airbnb/epoxy-ios/compare/171f63da...0.1.0) - 2021-02-01